% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcic.R
\name{mcic}
\alias{mcic}
\title{Modified correlation information criteria}
\usage{
mcic(object, tol = .Machine$double.eps, envir = parent.frame())
}
\arguments{
\item{object}{An object of class \code{"geem"} representing the fit.}

\item{tol}{A numeric value specifying the tolerance when calculating the
inverse of the model-based covariance matrix.
Default is \code{.Machine$double.eps}.}

\item{envir}{An environment object to which find the \code{"geem"} object.}
}
\value{
A numeric vector containing the mCIC value for the \code{"geem"}
object under consideration.
}
\description{
Calculating the modified correlation information criteria (mCIC)
based on Boonstra and Cavanaugh (in works).
}
\details{
mCIC is calculated based on the trace of the product of model-based
covariance matrices, where the referent variances are produced by the model
using the unstructured (i.e., general) correlation structure.
}
\examples{
data(ohio, package = "geepack")
fit <- geeM::geem(
  resp ~ age + smoke + age:smoke,
  id = id, data = ohio, family = stats::binomial(link = "logit"),
  corstr = "exchangeable"
)
geeM::mcic(fit)

}
