% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic.R
\name{vic}
\alias{vic}
\title{Variance information criteria}
\usage{
vic(object, robust = FALSE, envir = parent.frame())
}
\arguments{
\item{object}{An object of class \code{"geem"} representing the fit.}

\item{robust}{A logical denoting whether the robust covariance matrices
should be used. Default is \code{FALSE}.}

\item{envir}{An environment object to which find the \code{"geem"} object.
Default is \code{parent.frame()}.}
}
\value{
A numeric vector containing the VIC value for the \code{"geem"}
object under consideration.
}
\description{
Calculating the variance information criteria (VIC) based on
Boonstra and Cavanaugh (in works).
}
\details{
VIC is calculated based on the sum of the maximum of variance
ratios, where the referent variances are produced by the model using
the unstructured (i.e., general) correlation structure.

When the robust covariance matrices are used through VIC may be affected
by the variability and efficiency, while VIC based on the model-based
covariance matrices may be affected by bias. \emph{In the future, other robust
covariance estimators may be used.}
}
\examples{
data(ohio, package = "geepack")
fit <- geeM::geem(
  resp ~ age + smoke + age:smoke,
  id = id, data = ohio, family = stats::binomial(link = "logit"),
  corstr = "exchangeable"
)
geeM::vic(fit)

}
