% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic.R
\name{vic}
\alias{vic}
\title{Variance information criteria}
\usage{
vic(object, envir = parent.frame())
}
\arguments{
\item{object}{An object of class \code{"geem"} representing the fit.}

\item{envir}{An environment object to which find the \code{"geem"} object.}
}
\value{
A numeric vector containing the VIC value for the \code{"geem"}
object under consideration.
}
\description{
Calculating the variance information criteria (VIC) based on
Boonstra and Cavanaugh (in works).
}
\details{
VIC is calculated based on the sum of the maximum of model-based
variance ratios, where the referent variances are produced by the model using
the unstructured (i.e., general) correlation structure.
}
\examples{
data(ohio, package = "geepack")
fit <- geeM::geem(
  resp ~ age + smoke + age:smoke,
  id = id, data = ohio, family = stats::binomial(link = "logit"),
  corstr = "exchangeable"
)
geeM::vic(fit)

}
